move verify prepare/sample to eagle_utils helpers; dissolve EagleVerifyInputV2Mixin
